# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food-Delivery",
    "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
    
      "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
   "/api/mobile/cart/list": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/cart/list",
       "operationId": "index",
       "tags": ["Cart Listing"],
       "description": "get cart listing",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/cart/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/cart/add",
       "operationId": "index",
       "tags": ["Add to Cart"],
       "description": "Add Item to Cart",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "serviceId",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "orderPrice",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" },
         { "name": "orderTotalPrice",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "deliveryType",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "companyId",  "in": "formData",  "required":true,"type":"string"}


    ],

  
   
   
       "responses": {}
     }
   },
   "/api/mobile/cart/update": {
     "put": {
       "x-swagger-router-controller": "/api/mobile/cart/update",
       "operationId": "index",
       "tags": ["Update Cart"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "cartId",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "quantity",  "in": "formData",  "required":true ,"type":"string" }
         

    ],

  
   
   
       "responses": {}
     }
   },
...
   "/api/mobile/cart/remove": {
     "delete": {
       "x-swagger-router-controller": "/api/mobile/cart/remove",
       "operationId": "index",
       "tags": ["Remove Item from Cart"],
       "description": "get Prduct listing by Category ID",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true"},
         { "name": "cartId",  "in": "query",  "required":"true"}


    ],

 
       "responses": {}
     }
   },
   "/api/mobile/cart/clear": {
     "delete": {
       "x-swagger-router-controller": "/api/mobile/cart/clear",
       "operationId": "index",
       "tags": ["Clear Cart"],
       "description": "get Prduct listing by Category ID",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":"true"}


    ],

  
   
   
       "responses": {}
     }
   },
...
   "/api/mobile/tiffin/cart/add": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/add",
       "operationId": "index",
       "tags": ["Add Tiffin to cart "],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "tiffinId",  "in": "formData", "type":"string","required":true},
         { "name": "orderPrice",  "in": "formData", "type":"string","required":true},
         { "name": "orderTotalPrice",  "in": "formData", "type":"string","required":true},
         { "name": "companyId",  "in": "formData", "type":"string","required":true},
         { "name": "quantity",  "in": "formData", "type":"string","required":true},
         { "name": "fromDate",  "in": "formData", "type":"string","required":true},
         { "name": "package",  "in": "formData", "type":"string","required":true},
         { "name": "excPrice",  "in": "formData", "type":"string","description":"Excluded Total price"},
         { "name": "excAvaialbility",  "in": "formData", "type":"string","description":"Excluded Avaialbility if any like breakfast,lunch,dinner"},
         { "name": "excDays",  "in": "formData", "type":"string","description":"Excluded days like: monday,tuesday"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/cart/update": {
     "put": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/update",
       "operationId": "index",
       "tags": ["UPDATE Tiffin  cart "],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "cartId",  "in": "formData", "type":"string","required":true},
         { "name": "tiffinId",  "in": "formData", "type":"string","required":true},
         { "name": "orderPrice",  "in": "formData", "type":"string","required":true},
         { "name": "orderTotalPrice",  "in": "formData", "type":"string","required":true},
         { "name": "companyId",  "in": "formData", "type":"string","required":true},
         { "name": "quantity",  "in": "formData", "type":"string","required":true},
         { "name": "fromDate",  "in": "formData", "type":"string","required":true},
         { "name": "package",  "in": "formData", "type":"string","required":true},
         { "name": "excPrice",  "in": "formData", "type":"string"},
         { "name": "excAvaialbility",  "in": "formData", "type":"string"},
         { "name": "excDays",  "in": "formData", "type":"string"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/cart/list": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/list",
       "operationId": "index",
       "tags": ["Get Tiffin  cart  listing"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "lat",  "in": "query",  "required":true,"type":"string"},
         { "name": "lng",  "in": "query",  "required":true,"type":"string"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/cart/detail/{cartId}": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/list",
       "operationId": "index",
       "tags": ["Get Tiffin  cart  listing"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "cartId",  "in": "path",  "required":true,"type":"string"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/cart/remove": {
     "delete": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/remove",
       "operationId": "index",
       "tags": ["Remove Item From Cart"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "cartId",  "in": "query",  "required":true,"type":"string"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/cart/clear": {
     "delete": {
       "x-swagger-router-controller": "/api/mobile/tiffin/cart/clear",
       "operationId": "index",
       "tags": ["Clear Cart"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"}
        ],
       "responses": {}
     }
   },
...
  }
}